requirements.txt
requests~=2.32.3
streamlit~=1.39.0
pillow~=10.4.0
Faker~=30.3.0
google-cloud-storage~=2.18.2
python-dotenv~=1.0.1